+++ src/shared/index.ts
@@ -4,6 +4,14 @@ export * from './diff';
 export * from './types';
 export * from './constants';
 export * from './crdt';
-export { type DatasoleContract, type RpcParams, type EventData, type StateValue } from './contract';
+export {
+  type DatasoleContract,
+  type RpcParams,
+  type EventData,
+  type StateValue,
+  type RpcMethodKey,
+  type EventName,
+  type StateKeyName,
+} from './contract';
 export { type RpcResult as ContractRpcResult } from './contract';
 export { BUILD_CONSTANTS } from './build-constants';